feat: implements links to crates.io based on the package and version name
While reviewing the dependencies of a project, it is quite common to review the details of a certain package on crates.io. 

Graphviz can render hyperlinks on nodes and edges when the output is `svg`, `postscript` or `map`. 
This change adds a `href` property on the nodes in the dot file.
The url for the `href` is currently set to `https://crates.io/(?P<package>.*)(/(?P<version>.*))?`